Abandon custom builds and adopt new structure
A new structure to the main build system /mk was adopted, to make easier
to maintain and expand the ports. Custom builds were abandoned for the
same reason, the packages doesn't have a default behaviour and generate
a lot of files and variables during build phase (auto*hell) making the
task of using a custom and static build almost impossible (at least in a
portable way).