Use config.mk locally and use find(1) for pkgroot
Globally exporting the whole config.mk was messing with the auto*hell so
now the portsys_mk_* will import the config.mk again after saving the
old values.

Apparently assuming that pkgroot will be .pkgroot or build/.pkgroot does
not work, because the path is relative, so for now it will be using
find(1) (probably would be better to just use an absolute path).